commits ...

+++ database/migrations/2023_10_12_154563_create_customers_table.php
@@ -21,7 +21,7 @@ return new class extends Migration
             $table->string('photo')->nullable();
             $table->string('marital_status');
             $table->integer('phone');
-            $table->integer('nida');
+            $table->bigInteger('nida');
             $table->string('region');
             $table->string('district');
             $table->string('street');
